New discount dates, enhanced goals table
Lodges with a name but not a name in the table are attributed to “At
Large” so the count’s still solid. Tweaked the way rows are calculated
to compensate (need to set initial goals to 0 in the table). Also
switched to using jQuery’s text() instead of html() for parsing
integers, just in case.